Readme.txt
A leading bank wants to develop a customer segmentation to give promotional offers to its customers. They collected a sample that summarizes the activities of users during the past few months. It is a summarized data that contains the average values in all the columns considering all the months, and not for any particular month. The task is to identify the segments based on credit card usage.